requirements.txt
google-generativeai
langchain
langchain-google-genai
langchain-core
langchain-community
torch
torchvision
pillow
python-dotenv